Feat: Added a custom hook and refactored code

+++ src/context/videoContext.js
@@ -14,47 +14,58 @@ function VideoProvider({ children }) {
 
 
     const likeHandler = async (video) => {
-        try {
-            const response = await axios.post(
-                "/api/user/likes",
-                { video },
-                {
-                    headers: {
-                        authorization: encodedToken,
-                    },
+        if (encodedToken) {
+            try {
+                const response = await axios.post(
+                    "/api/user/likes",
+                    { video },
+                    {
+                        headers: {
+                            authorization: encodedToken,
+                        },
+                    }
+                );
+                if (response.status === 201) {
+                    videoDispatch({
+                        type: "SET_LIKED_VIDEOS",
+                        payload: { likedVideos: response.data.likes },
+                    });
+                    toast.success("Added to Liked Videos");
                 }
-            );
-            if (response.status === 201) {
-                videoDispatch({
-                    type: "SET_LIKED_VIDEOS",
-                    payload: { likedVideos: response.data.likes },
-                });
-                toast.success("Added to Liked Videos");
+            } catch (err) {
+                console.log(err);
+                toast.error(err.response.data.errors[0])
             }
-        } catch (err) {
-            console.log(err);
-            toast.error(err.response.data.errors[0])
+        }
+        else {
+            toast.info("Please Login");
         }
     };
 
     const unLikeHandler = async (video) => {
-        try {
-            const response = await axios.delete(`/api/user/likes/${video._id}`, {
-                headers: {
-                    authorization: encodedToken,
-                },
-            });
-            if (response.status === 200) {
-                videoDispatch({
-                    type: "SET_LIKED_VIDEOS",
-                    payload: { likedVideos: response.data.likes },
+        if (encodedToken) {
+            try {
+                const response = await axios.delete(`/api/user/likes/${video._id}`, {
+                    headers: {
+                        authorization: encodedToken,
+                    },
                 });
-                toast.error("Removed from Liked Videos");
+                if (response.status === 200) {
+                    videoDispatch({
+                        type: "SET_LIKED_VIDEOS",
+                        payload: { likedVideos: response.data.likes },
+                    });
+                    toast.error("Removed from Liked Videos");
+                }
+            } catch (err) {
+                console.log(err);
+                toast.error(err.response.data.errors[0])
             }
-        } catch (err) {
-            console.log(err);
-            toast.error(err.response.data.errors[0])
         }
+        else {
+            toast.info("Please Login")
+        }
+
     };
 
     const addToWatchLaterHandler = async (video) => {
@@ -88,26 +99,32 @@ function VideoProvider({ children }) {
     }
 
     const removeFromWatchlaterHandler = async (video) => {
-        try {
-            const response = await axios.delete(`/api/user/watchlater/${video._id}`,
-                {
-                    headers: {
-                        authorization: encodedToken
+        if (encodedToken) {
+            try {
+                const response = await axios.delete(`/api/user/watchlater/${video._id}`,
+                    {
+                        headers: {
+                            authorization: encodedToken
+                        }
                     }
+                )
+                if (response.status === 200) {
+                    videoDispatch({
+                        type: "SET_WATCHLATER_VIDEOS",
+                        payload: { watchLaterVideos: response.data.watchlater },
+                    });
+                    toast.error("Removed from Watch Later")
                 }
-            )
-            if (response.status === 200) {
-                videoDispatch({
-                    type: "SET_WATCHLATER_VIDEOS",
-                    payload: { watchLaterVideos: response.data.watchlater },
-                });
-                toast.error("Removed from Watch Later")
+            }
+            catch (err) {
+                console.log(err)
+                toast.error(err.response.data.errors[0])
             }
         }
-        catch (err) {
-            console.log(err)
-            toast.error(err.response.data.errors[0])
+        else {
+            toast.info("Please Login")
         }
+
     }
 
     const removeFromHistoryHandler = async (video) => {
